squash: create vagabond state server
