+++ server/Patches/ProfileBootstrapPatch.cs
@@ -31,7 +31,7 @@ public sealed class ProfileBootstrapPatch : AbstractPatch
                 return;
             }
 
-            var state = VagabondService.GetState(sessionId);
+            var state = VagabondStateService.GetState(sessionId);
             if (!state.VagabondModeEnabled)
             {
                 return;
@@ -42,7 +42,7 @@ public sealed class ProfileBootstrapPatch : AbstractPatch
             if (state.ResetProfile)
             {
                 state.ResetProfile = false;
-                VagabondService.SaveState(sessionId, state);
+                VagabondStateService.SaveState(sessionId, state);
                 VirtualStashService.ClearAllTraderStashes(sessionId);
                 VagabondService.ResetProfile(sessionId, pmc);
                 VagabondService.PersistProfileIfPossible(sessionId);
