+++ server/Routes/VagabondRouter.cs
@@ -69,7 +69,7 @@ public class VagabondRouter(
             return response;
         }
 
-        var state = VagabondService.GetState(sessionId);
+        var state = VagabondStateService.GetState(sessionId);
         // load their hideout first time
         if (ExfilService.AddHideoutExfil(pmc.CharacterData.PmcData, state))
         {
@@ -121,7 +121,7 @@ public class VagabondRouter(
             return response;
         }
 
-        var state = VagabondService.GetState(sessionId);
+        var state = VagabondStateService.GetState(sessionId);
 
         if (state.HideoutState != null && (!VagabondConfig.Config.AllowHideoutRelocation && !state.CanPlaceHideout))
         {
@@ -156,7 +156,7 @@ public class VagabondRouter(
         ExfilService.AddHideoutExfil(pmc.CharacterData.PmcData, state);
         ExfilService.BuildCustomExfilSnapshot(true);
 
-        VagabondService.SaveState(sessionId, state);
+        VagabondStateService.SaveState(sessionId, state);
         response.Success = true;
         response.CurrentRaid = mapName;
         response.MapName = mapName;
